32b to 64b via cursor
Also switch from BIOS to UEFI. BIOS is legacy and boots in real mode
(16b) and then has to be switched to protected mode (32b) or long
mode (64b). UEFI is less legacy and has easier support for 64b.

Both are firmware that pass off control to the bootloader before
the bootloader starts the kernel.

BIOS supports VGA, UEFI does not, so we lost our "hello world" print
in this commit. We need a framebuffer in order to print again.

Note we still need to do the GDT but we'll get there.

Also we switched to multiboot2 here because it supports UEFI.